services/ingestion/requirements.txt
redis==5.0.1
praw==7.7.1
requests==2.31.0
python-dotenv==1.0.0
alpaca-py==0.14.0
newsapi-python==0.2.7
